+++ projects/komponentkartan/src/lib/controls/input/input.component.ts
@@ -101,7 +101,6 @@ export class InputComponent implements ControlValueAccessor, OnChanges, OnInit {
   onBlur(event) {
     this.onTouched();
     this.hasFocus = false;
-    this.blur.emit(event);
   }
 
   onFocus(event) {