+++ projects/komponentkartan/src/lib/controls/menu/menu-item.component.ts
@@ -1,13 +1,13 @@
 import { Input, Component, HostListener, ElementRef, forwardRef, HostBinding, AfterViewInit, ViewChild, Renderer2 } from '@angular/core';
-import { Router, RouterLinkActive } from '@angular/router';
-import { MenuItemBase } from './menu-item-base';
+import { Router } from '@angular/router';
+import { MenuItemBaseDirective } from './menu-item-base';
 
 @Component({
     selector: 'vgr-menu-item',
     templateUrl: './menu-item.component.html',
-    providers: [{ provide: MenuItemBase, useExisting: forwardRef(() => MenuItemComponent) }]
+    providers: [{ provide: MenuItemBaseDirective, useExisting: forwardRef(() => MenuItemComponent) }]
 })
-export class MenuItemComponent extends MenuItemBase implements AfterViewInit {
+export class MenuItemComponent extends MenuItemBaseDirective implements AfterViewInit {
     @Input() text: string;
     @Input() link: string;
     @Input() disabled = false;
@@ -18,7 +18,7 @@ export class MenuItemComponent extends MenuItemBase implements AfterViewInit {
     @Input() isInternalLink = true;
     @HostBinding('attr.role') role = 'menuitem';
     @HostBinding('attr.aria-disabled') ariaDisabled;
-    @ViewChild('menuitem', { static: false }) menuitem: ElementRef;
+    @ViewChild('menuitem') menuitem: ElementRef;
 
     @HostListener('keydown', ['$event']) onKeyDown(event: KeyboardEvent) {