+++ src/app/dropdown-select-documentation/dropdown-select-documentation.component.ts
@@ -1,18 +1,21 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { FormControl } from '@angular/forms';
+import { DropdownSelectComponent } from '../../../projects/komponentkartan/src/lib';
 
 @Component({
   selector: 'app-dropdown-select-documentation',
   templateUrl: './dropdown-select-documentation.component.html'
 })
 export class DropdownSelectDocumentationComponent {
-
+  @ViewChild('myDropdown', { read: DropdownSelectComponent }) myDropdown: DropdownSelectComponent;
   items5: string[];
   items10: string[];
   items50: string[];
   form: FormControl;
   itemLabel = 'Ett långt alternativ som skrivs ut i helhet';
   itemValue = 1;
+  readonly = false;
+  disabled = false;
 
   constructor() {
     this.items5 = this.getItems(5);
@@ -27,6 +30,18 @@ export class DropdownSelectDocumentationComponent {
       items.push(`Val ${item}`);
     }
     return items;
+  }
+
+  setFocus() {
+    this.myDropdown.focus();
+  }
+
+  toggleReadonly() {
+    this.readonly = !this.readonly;
+  }
+
+  toggleDisabled(value) {
+    this.disabled = !this.disabled;
   }
 
 }